third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/block-container-block-end-expected.txt

This is a testharness.js-based test.
FAIL item 1 assert_equals: 
<item data-expected-margin-bottom="0"></item>
margin-bottom expected "0" but got "10"
Harness: the test ran to completion.
